Improve intent of policy validation code
Create domain notion for result that have no validation errors instead
of using null references.  See ValidationResult.noErrors .  Would really
like to enforce that ValidationResult is not null but thats beyond the
scope of current work.

Make AbstractActionType return a non-null ValidationResult, reducing the
magic of null.

Rename PolicyTest to PolicyValidationTest since this is what it's
focused on and I wouldn't want it to include tests of other policy
features.

PR #234
Done during CLMSUPPORT-50 but not specific to it.