Create a new DW service per functional test class
Create one DW service per test class execution so that it properly
starts up on every execution.  There's a bug in the DW rule that assumes
it's destroyed after every execution (as would be the case if this
wasn't shared in the abstract base class).

Incidentally this is what the Geb/Spock tests do, as discovered through
debugging.

Now have a problem where sometimes the webdriver hangs on the next test
class waiting for a server response.  If I manually refresh the page the
driver loaded it works, but that's not scalable =)

JUnitTests is only for debugging.  Will remove once all problems are
addressed.