INT-1634 new '--metadata-file' param (#3971)
A new parameter '--metadata-file' was added to the CLI to allow the user to pass in extra information with the request.  In this case the metadata file only contains a commit hash attribute, which will eventually be used by downstream processes to annotate GitHub commits and pull requests.  In the future this metadata will be enhanced with other information.