Test to remind us in case ReactTextInput is not needed anymore (#4820)
Introducing SeleniumReactInteractionTest which will fail once Selenium starts working with React fields properly. This test will fail once that is the case and it will inform us that we do not need the workarounds implemented in ReactTextInput. This class also comes with introductions on what to do when that day comes.