SDEV-373: Upgrade React Shared Components (#9869)
Upgrading RSC to v12.9.0 will remove noise from the logs by removing a warning. This PR is also a first step in making a path forward for future RSC upgrades which have been stalled due to problem with NXRM.

To work around the NXRM impasse the one component is being install directly from npm. The risk should be quite minimal sense this is our own component that we publish. We are only bypassing npm for this coordinate.