SDEV-337: Optimizing Application without Automated Source Control Feedback (#9874)
This spike has determined that some performance optimization is needed to get fetch time for apps without automated source control feedback down to an acceptable level

One of the biggest optimizations is changing the API from true pagination to simply taking a limit. We don't actually make use of pagination in the UI in current designs. By using a limit instead of pagination we don't need to compute the full set of apps without scm feed back enabled. Once we have enough to satisfy the limit we can stop. For fetching 10 apps in cases where lots of apps have this functionality disabled the fetch now often takes less than a second.

The worst case scenario for this new approach is when all apps have the functionality enabled. In that case it will still have to compute the full set as before. Some minor optimizations have spead this up a little.

Testing on local with 10 runs against each algorithm, a fresh system start before each set of runs found that the current algorithm was taking

An average of 55.9261 seconds running against 40K apps

For the new algorithm it is taking only 25.71015 seconds on average

Note: That tests reflects ideals, the original algorithm was observed taking as much as 2 minutes with other applications open. This type of testing can not tell us exactly how long these operations will take on a true production setup but do indicate progress has been made.

Some of the other minor optimizations that help in the full list walk were:

Making sure we only have to fetch apps from the db once
Making sure we only have to compute the composite SourceControl component once per apps

Co-authored-by: Anastasia Van Riet <110420624+anastasia-vanriet@users.noreply.github.com>