SDEV-475: Part I -- Use License Flag instead of Feature flag (#10052)
We would like to be able to enable the Developer Dasbhoard on our end for all PPP customers. The feature flag was not allowing us to do this. This switches the frontend over from using the feature flag to using the new license feature for developer dashboard.

A follow on PR will remove the old feature flag from IQ