Move Integrations Tests To Jest (#10018)
Now that we have the ability to write tests in jest we should start migrating existing ones. Once we get all tests moved we can remove jasmine which is no longer well supported and slows down our builds.

This is another baby step in that direction. It converts all tests for the integrations module from jasmine to jest. Mostly this is just a matter of renaming the files. A few small tweaks were needed