SDEV-690 - Add Endpoint To Tell Frontend If User Should See Charts (#10750)
Going forward we would like to conditionally hide and show content on the Developer Dashboard page depending on the user. The first pass at this will be to hide the change over time metrics from folks who don't have Read on Root.

To support this we are adding an endpoint that can return information to the frontend about what should be shown or hidden on the page.