SDEV-900: Restore Development Dashboard Metric Collection (#11102)
The development dashboard metric collection was disabled following #incident-236. This change restores the metric collection in a more performant way.

The main problem was that to calculate mean time to remediate we pulled the entire policy_violation table into memory. This could lead to an Out of Memory exception which crashed the IQ Server when the this table is large.

To avoid this we now:

    * compute the average completely in the database, so definitely no OOM
    * use a ~3 month sliding window for the average to reduce db compute load
    * add an index on open_time so that the sliding window does not cause a sequential scan

As an extra precaution, I've added a feature flag . It will be on by default, but if there is another incident like 236, we can help affected customers disable the feature for their instance.