SDEV-1029: Don't Drop Alerts When There is No Re-Eval (#11176)
Fixes a bug where the api was returning an empty list of alerts if no changes were detected to the labels

There is some chance that "pathname in the component facts inside the alerts" may still be dropped. This will need to be looked at in more detail by the team who owns this module. In the meantime, this one liner, fixes the most blocking aspect of this bug