SDEV-1158: Change Criteria for CICD Integrated Status (#11497)
A customer using gitlab has complained that despite near total cicd integration the developer dashboard
shows them as having integrated a very small percentage of their apps. This is because we have been relying on scanTriggerType (CONTINUOUS_INTEGRATION). As discussed this is not reliable, the gitlab plugin will not send this. We discussed

Fixing the gitlab plugin to correctly report its self, but it was determined that a more accurate reflection of integration was
regular build stage scans. The criteria is that the app has as at least one build stage scan in the last 3 months.

To count how many apps were integrated at a past point in time we will check the number of apps with build stage evaluations within the last 3 months from that point in time.

We discount evals older than 3 months because apps integrated with ci/cd should be evaluated regularly. If they were evaluated once and then not again it could have simply been someone triggering a build stage eval from the cli on local or testing it out with file uploads. If it's truly integrated it will keep getting evaluated.

This should solve this customers issue and address a wider range of possible customer ci/cd configurations

Co-authored-by: Anastasia Van Riet <110420624+anastasia-vanriet@users.noreply.github.com>