CLM-20811: Phese-III Re-Eval -- UI reeval prevention (#11826)
Several recent issues reported have been attributed to the fact that we do not save data in the policy_violation_table when re-evaluation a scan that is not the latest for that stage. There are good reasons for not persisting these violations, but unfortunately this leaves the report in bad state where some functionality is broken. To prevent this we have now blocked re-evals on non-latests scans. This has already been done on the backend.

This work follows on that and tries to improve the UX by disabling the re-eval button when we know this is not the latest report and offering friendly messaging with a link to the latest report.
Changes

    If a user is viewing a report that is not the latest for the stage we will:
        disable the re-evaluate button
        show a banner explaining with a link to the newer report (except on develop stage reports)
        There is a tooltip on the disabled button explaining why it’s disabled